Remove all magic strings, introduce strongly typed object for Configuration
In .NET Core, you can create a POCO that will be bound to some
configuration value in appsettings.json. This is really nice to avoid
magic strings and other duplication and passing of config.

Once this was setup, used the new `@inject` to inject the configuration
into every View in the `_Layout`. I chose this approach because it
allows easy access to "global" config.